[DRUP-524] Refine auto assignment legal name tests.
